SAK-17586
git-svn-id: https://source.sakaiproject.org/svn/authz/trunk@69855 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ authz/authz-tool/tool/src/java/org/sakaiproject/authz/tool/RealmsAction.java
@@ -201,8 +201,12 @@ public class RealmsAction extends PagedResourceActionII
 		// put all realms into the context
 		context.put("realms", realms);
 
-		int pageSize = Integer.valueOf(state.getAttribute(STATE_PAGESIZE).toString()).intValue();
-		int currentPageNubmer = Integer.valueOf(state.getAttribute(STATE_CURRENT_PAGE).toString()).intValue();
+		// defaults to page size of 20
+		int pageSize = state.getAttribute(STATE_PAGESIZE) != null ? Integer.valueOf(state.getAttribute(STATE_PAGESIZE).toString()).intValue() : 20;
+		
+		// defaults to page 1
+		int currentPageNubmer = state.getAttribute(STATE_CURRENT_PAGE) != null ? Integer.valueOf(state.getAttribute(STATE_CURRENT_PAGE).toString()).intValue() : 1;
+		
 		int startNumber = pageSize * (currentPageNubmer - 1) + 1;
 		int endNumber = pageSize * currentPageNubmer;
 
@@ -289,7 +293,7 @@ public class RealmsAction extends PagedResourceActionII
 		// build the menu
 		// we need the form fields for the remove...
 		Menu bar = new MenuImpl();
-		if (AuthzGroupService.allowRemove(realm.getId()))
+		if (realm != null && AuthzGroupService.allowRemove(realm.getId()))
 		{
 			bar.add(new MenuEntry(rb.getString("realm.remove"), null, true, MenuItem.CHECKED_NA, "doRemove", "realm-form"));
 		}
@@ -353,7 +357,8 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get all roles
 		List allRoles = new Vector();
-		allRoles.addAll(realm.getRoles());
+		if (realm != null)
+			allRoles.addAll(realm.getRoles());
 		Collections.sort(allRoles);
 		context.put("allRoles", allRoles);
 
@@ -384,7 +389,8 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get all roles
 		List allRoles = new Vector();
-		allRoles.addAll(realm.getRoles());
+		if (realm != null)
+			allRoles.addAll(realm.getRoles());
 		Collections.sort(allRoles);
 		context.put("allRoles", allRoles);
 
@@ -412,7 +418,8 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get all roles
 		List allRoles = new Vector();
-		allRoles.addAll(realm.getRoles());
+		if (realm != null)
+			allRoles.addAll(realm.getRoles());
 		Collections.sort(allRoles);
 		context.put("allRoles", allRoles);
 
@@ -435,16 +442,20 @@ public class RealmsAction extends PagedResourceActionII
 		context.put("user", user);
 
 		// get this user's role - only if not provided
-		Member grant = realm.getMember(user.getId());
-		context.put("grant", grant);
-		if ((grant != null) && (!grant.isProvided()) && (grant.getRole() != null))
+		if (realm != null)
 		{
-			context.put("roles", grant.getRole());
+			Member grant = realm.getMember(user.getId());
+			context.put("grant", grant);
+			if ((grant != null) && (!grant.isProvided()) && (grant.getRole() != null))
+			{
+				context.put("roles", grant.getRole());
+			}
 		}
 
 		// get all roles
 		List allRoles = new Vector();
-		allRoles.addAll(realm.getRoles());
+		if (realm != null)
+			allRoles.addAll(realm.getRoles());
 		Collections.sort(allRoles);
 		context.put("allRoles", allRoles);
 
@@ -715,25 +726,27 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get the realm
 		AuthzGroup realm = (AuthzGroup) state.getAttribute("realm");
-
-		// remove the realm
-		try
+		if (realm != null)
 		{
-			AuthzGroupService.removeAuthzGroup(realm);
+			// remove the realm
+			try
+			{
+				AuthzGroupService.removeAuthzGroup(realm);
+			}
+			catch (AuthzPermissionException e)
+			{
+				addAlert(state, rb.getString("realm.notpermis2") + " " + realm.getId());
+			}
+	
+			// cleanup
+			cleanState(state);
+	
+			// go to main mode
+			state.removeAttribute("mode");
+	
+			// make sure auto-updates are enabled
+			enableObserver(state);
 		}
-		catch (AuthzPermissionException e)
-		{
-			addAlert(state, rb.getString("realm.notpermis2") + " " + realm.getId());
-		}
-
-		// cleanup
-		cleanState(state);
-
-		// go to main mode
-		state.removeAttribute("mode");
-
-		// make sure auto-updates are enabled
-		enableObserver(state);
 
 	} // doRemove_confirmed
 
@@ -864,8 +877,11 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get the role
 		AuthzGroup realm = (AuthzGroup) state.getAttribute("realm");
-		Role role = realm.getRole(id);
-		state.setAttribute("role", role);
+		if (realm != null)
+		{
+			Role role = realm.getRole(id);
+			state.setAttribute("role", role);
+		}
 
 	} // doEdit_role
 
@@ -878,15 +894,17 @@ public class RealmsAction extends PagedResourceActionII
 
 		AuthzGroup realm = (AuthzGroup) state.getAttribute("realm");
 		Role role = (Role) state.getAttribute("role");
-
-		// remove the role (no confirm)
-		realm.removeRole(role.getId());
-
-		// done with the role
-		state.removeAttribute("role");
-
-		// return to edit mode
-		state.setAttribute("mode", "edit");
+		if (realm != null && role != null)
+		{
+			// remove the role (no confirm)
+			realm.removeRole(role.getId());
+	
+			// done with the role
+			state.removeAttribute("role");
+	
+			// return to edit mode
+			state.setAttribute("mode", "edit");
+		}
 
 	} // doRemove_role
 
@@ -985,7 +1003,7 @@ public class RealmsAction extends PagedResourceActionII
 
 		// we are setting for either a new role or this role
 		Role role = (Role) state.getAttribute("role");
-		if (role == null)
+		if (realm != null && role == null)
 		{
 			// read the form
 			String id = StringUtil.trimToNull(data.getParameters().getString("id"));
@@ -1058,15 +1076,18 @@ public class RealmsAction extends PagedResourceActionII
 			return false;
 		}
 
-		// create the role
-		try
+		if (realm != null)
 		{
-			realm.addRole(id, role);
-		}
-		catch (RoleAlreadyDefinedException e)
-		{
-			addAlert(state, rb.getString("realm.arole") + id + rb.getString("realm.defined"));
-			return false;
+			// create the role
+			try
+			{
+				realm.addRole(id, role);
+			}
+			catch (RoleAlreadyDefinedException e)
+			{
+				addAlert(state, rb.getString("realm.arole") + id + rb.getString("realm.defined"));
+				return false;
+			}
 		}
 
 		return true;
@@ -1127,15 +1148,18 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get the realm
 		AuthzGroup realm = (AuthzGroup) state.getAttribute("realm");
-
-		// clear out this user's settings
-		realm.removeMember(user.getId());
-
-		// done with the user
-		state.removeAttribute("user");
-
-		// return to edit mode
-		state.setAttribute("mode", "edit");
+		
+		if (realm != null && user != null)
+		{
+			// clear out this user's settings
+			realm.removeMember(user.getId());
+	
+			// done with the user
+			state.removeAttribute("user");
+	
+			// return to edit mode
+			state.setAttribute("mode", "edit");
+		}
 
 	} // doRemove_user
 
@@ -1212,26 +1236,28 @@ public class RealmsAction extends PagedResourceActionII
 
 		// get the realm
 		AuthzGroup realm = (AuthzGroup) state.getAttribute("realm");
-
-		// if the user is set to have the same role the user already has, do nothing
-		Member grant = realm.getMember(user.getId());
-
-		// if no change, change nothing
-		if (roles == null && (grant == null && (checkForStatus == null) ) ) {
-		    // removing this since it would cause a null pointer exception if the code got here so we think the code does not ever get here -AZ
-		    //) || (grant.isProvided())))) return true;
-		    return true;
-		}
-		if ((roles != null) && (grant != null) && (grant.getRole().getId().equals(roles) && (grant.isActive()==status)) && !grant.isProvided()) return true;
-
-		// clear out this user's settings
-		realm.removeMember(user.getId());
-
-		// if there's a role, give it
-		if (roles != null)
+		if (realm != null)
 		{
-			// TODO: active, provided
-			realm.addMember(user.getId(), roles, status, false);
+			// if the user is set to have the same role the user already has, do nothing
+			Member grant = realm.getMember(user.getId());
+	
+			// if no change, change nothing
+			if (roles == null && (grant == null && (checkForStatus == null) ) ) {
+			    // removing this since it would cause a null pointer exception if the code got here so we think the code does not ever get here -AZ
+			    //) || (grant.isProvided())))) return true;
+			    return true;
+			}
+			if ((roles != null) && (grant != null) && (grant.getRole().getId().equals(roles) && (grant.isActive()==status)) && !grant.isProvided()) return true;
+	
+			// clear out this user's settings
+			realm.removeMember(user.getId());
+	
+			// if there's a role, give it
+			if (roles != null)
+			{
+				// TODO: active, provided
+				realm.addMember(user.getId(), roles, status, false);
+			}
 		}
 
 		return true;
